Move Topbar into src/ui and strip stylesheet comments
- Relocate the Topbar component (Topbar, Brand, MobileOverflow) from
  src/components to src/ui; fix the affected relative imports.
- Remove all comments from the topbar/ui StyledWrapper stylesheets.
